# desktop/src/open-api/model/supplierProfile.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';
import { SupplierProfileAlias } from './supplierProfileAlias';
import { Tag } from './tag';


/**
 * Group-scoped supplier profile that stores optional receipt-review defaults. A profile is a suggestion, never an automatic classification.
 */
export interface SupplierProfile { 
    id?: number;
    createdAt?: string;
    createdBy?: number;
    createdByString?: string;
    updatedAt?: string;
    groupId?: number;
    /**
     * Canonical supplier display name
     */
    name?: string;
    /**
     * Normalised canonical name used for matching
     */
    normalisedName?: string;
    /**
     * Optional expected ISO 4217 document currency
     */
    expectedDocumentCurrencyCode?: string;
    /**
     * Disabled profiles remain stored but do not match
     */
    enabled?: boolean;
    categories?: Array<Category>;
    tags?: Array<Tag>;
    aliases?: Array<SupplierProfileAlias>;
}