# desktop/src/open-api/model/supplierProfileAlias.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SupplierProfileAlias { 
    id?: number;
    createdAt?: string;
    createdBy?: number;
    createdByString?: string;
    updatedAt?: string;
    supplierProfileId?: number;
    groupId?: number;
    /**
     * Display alias as entered
     */
    name?: string;
    /**
     * Normalised alias used for matching
     */
    normalisedName?: string;
}