# desktop/src/open-api/model/upsertSupplierProfileCommand.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpsertSupplierProfileCommand { 
    /**
     * Canonical supplier display name
     */
    name: string;
    /**
     * Additional names that should match this profile
     */
    aliases?: Array<string>;
    /**
     * Default category ids from the catalogue
     */
    categoryIds?: Array<number>;
    /**
     * Default tag ids from the catalogue
     */
    tagIds?: Array<number>;
    /**
     * Optional expected ISO 4217 document currency
     */
    expectedDocumentCurrencyCode?: string;
    /**
     * Whether the profile participates in matching. Defaults to true on create.
     */
    enabled?: boolean;
}